+++ HISTORY.rst
@@ -6,6 +6,9 @@ Pending
 -------
 
 * Fix modernize running on Python 2.
+* Run ``isort`` in the same Python process rather than with ``subprocess``
+* Properly gate ``flake8`` and ``isort`` so that they run only if they are
+  installed.
 
 1.0.0 (2016-06-19)
 ------------------